refactor: 요청/응답에서 공통으로 사용하는 객체들 패키지 이동
