+++ src/test/java/webserver/http/request/RequestLineTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import webserver.http.protocol.Protocol;
 
 import java.util.HashMap;
 import java.util.Map;
