+++ src/test/java/webserver/http/request/parser/RequestLineParserTest.java
@@ -3,8 +3,9 @@ package webserver.http.request.parser;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import webserver.http.protocol.parser.ProtocolParser;
 import webserver.http.request.Method;
-import webserver.http.request.Protocol;
+import webserver.http.protocol.Protocol;
 import webserver.http.request.RequestLine;
 
 import static org.assertj.core.api.Assertions.assertThat;
